Refactored for multiple use cases.
